+++ src/HeifChromaDownsamplingAlgorithm.cs
@@ -1,0 +1,49 @@
+﻿/*
+ * .NET bindings for libheif.
+ * Copyright (c) 2020, 2021, 2022, 2023 Nicholas Hayes
+ *
+ * Portions Copyright (c) 2017 struktur AG, Dirk Farin <farin@struktur.de>
+ *
+ * This file is part of libheif-sharp.
+ *
+ * libheif-sharp is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of
+ * the License, or (at your option) any later version.
+ *
+ * libheif-sharp is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libheif-sharp.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace LibHeifSharp
+{
+    /// <summary>
+    /// The LibHeif chroma down-sampling algorithms.
+    /// </summary>
+    /// <seealso cref="HeifColorConversionOptions"/>
+    public enum HeifChromaDownsamplingAlgorithm
+    {
+        /// <summary>
+        /// The nearest neighbor chroma down-sampling algorithm.
+        /// </summary>
+        NearestNeighbor = 1,
+
+        /// <summary>
+        /// The average chroma down-sampling algorithm.
+        /// </summary>
+        Average = 2,
+
+        /// <summary>
+        /// The sharp YUV chroma down-sampling algorithm.
+        /// </summary>
+        /// <remarks>
+        /// Makes edges look sharper when using YUV 420 with bilinear chroma up-sampling.
+        /// </remarks>
+        SharpYuv = 3,
+    }
+}