+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Added
+
+* A `ColorConversionOptions` property to the `HeifDecodingOptions` class.
+  * Requires LibHeif 1.16.0 or later, it will be ignored on older versions.
+
 ## Fixed
 
 * The `PixelAspectRatio` property will ignore a default initialized structure.